Added a convolution model based on Kolmogorov-Arnold theory
The material is taken from external sources, my merit is that it turned out to be combined. The initial results are very positive - the learning rate and accuracy are significantly higher. However, the model weighs more, we need to figure it out further.

the example is taken from https://github.com/AntonioTepsich/Convolutional-KANs.git
Link to the article https://arxiv.org/pdf/2406.13155