The markup function has been upgraded
The markup process is running. The current neuromodel does not filter out unnecessary things well.
It will be necessary to add waveforms without events and/or expand the area "around the event".
Probably, the marking of the "switching" events will have to be changed, because now they are very long in most cases, and the process there is very smooth (as can be seen on the entire waveform, not one period)