Refactoring
The normalization function has been moved to the appropriate class.

At the same time, it is worthwhile to resolve the issue of different "__init__" in the future. Since now, these classes were doing slightly different tasks.

Before that, normalization was LOOKING for coefficients, and now normalization of the existing waveform is based on the coefficients found.