The first attempt at neuro PDR training
This is the first launch of neuro PDR training. There may be a lot of jambs here, while it is of the utmost importance to arrange the launch itself and the passage of at least one cycle. The rest will be adjusted later.

Alas, for now it is necessary to fix the jambs in the dataset.