Upgrading the merge_normalization_files function
The _get_max_bus_number_from_columns function has also been redesigned, now it can be called independently and is not tied to the creation of a class. In general, it will be necessary to rework the entire class, because setting the folder where to look is definitely not necessary at the beginning, it causes a lot of problems.